Fix event loop.
Shading does not seem to work.

+++ src/resource/mesh.rs
@@ -16,7 +16,7 @@ mod error;
 /// It also contains the GPU location of those buffers.
 pub struct Mesh {
     coords: Arc<RwLock<GPUVec<Point3<f32>>>>,
-    faces: Arc<RwLock<GPUVec<Point3<i32>>>>,
+    faces: Arc<RwLock<GPUVec<Point3<u16>>>>,
     normals: Arc<RwLock<GPUVec<Vector3<f32>>>>,
     uvs: Arc<RwLock<GPUVec<Point2<f32>>>>,
 }
@@ -27,7 +27,7 @@ impl Mesh {
     /// If the normals and uvs are not given, they are automatically computed.
     pub fn new(
         coords: Vec<Point3<f32>>,
-        faces: Vec<Point3<i32>>,
+        faces: Vec<Point3<u16>>,
         normals: Option<Vec<Vector3<f32>>>,
         uvs: Option<Vec<Point2<f32>>>,
         dynamic_draw: bool,
@@ -147,7 +147,7 @@ impl Mesh {
     /// Creates a new mesh. Arguments set to `None` are automatically computed.
     pub fn new_with_gpu_vectors(
         coords: Arc<RwLock<GPUVec<Point3<f32>>>>,
-        faces: Arc<RwLock<GPUVec<Point3<i32>>>>,
+        faces: Arc<RwLock<GPUVec<Point3<u16>>>>,
         normals: Arc<RwLock<GPUVec<Vector3<f32>>>>,
         uvs: Arc<RwLock<GPUVec<Point2<f32>>>>,
     ) -> Mesh {
@@ -215,7 +215,7 @@ impl Mesh {
     }
 
     /// This mesh faces.
-    pub fn faces(&self) -> &Arc<RwLock<GPUVec<Point3<i32>>>> {
+    pub fn faces(&self) -> &Arc<RwLock<GPUVec<Point3<u16>>>> {
         &self.faces
     }
 
@@ -237,7 +237,7 @@ impl Mesh {
     /// Computes normals from a set of faces.
     pub fn compute_normals_array(
         coordinates: &[Point3<f32>],
-        faces: &[Point3<i32>],
+        faces: &[Point3<u16>],
     ) -> Vec<Vector3<f32>> {
         let mut res = Vec::new();
 
@@ -249,7 +249,7 @@ impl Mesh {
     /// Computes normals from a set of faces.
     pub fn compute_normals(
         coordinates: &[Point3<f32>],
-        faces: &[Point3<i32>],
+        faces: &[Point3<u16>],
         normals: &mut Vec<Vector3<f32>>,
     ) {
         let mut divisor: Vec<f32> = iter::repeat(0f32).take(coordinates.len()).collect();
