+++ src/scene/object.rs
@@ -266,7 +266,7 @@ impl Object {
 
     /// Mutably access the object's faces.
     #[inline(always)]
-    pub fn modify_faces<F: FnMut(&mut Vec<Point3<i32>>)>(&mut self, f: &mut F) {
+    pub fn modify_faces<F: FnMut(&mut Vec<Point3<u16>>)>(&mut self, f: &mut F) {
         let bmesh = self.mesh.borrow_mut();
         let _ = bmesh
             .faces()
@@ -279,7 +279,7 @@ impl Object {
 
     /// Access the object's faces.
     #[inline(always)]
-    pub fn read_faces<F: FnMut(&[Point3<i32>])>(&self, f: &mut F) {
+    pub fn read_faces<F: FnMut(&[Point3<u16>])>(&self, f: &mut F) {
         let bmesh = self.mesh.borrow();
         let _ = bmesh
             .faces()
